move contactSchema to services
